server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==1.10.18
httpx==0.27.2
numpy==2.1.0
Pillow==10.4.0
grpcio==1.62.2
protobuf==5.28.2
